Refactor MeasurementDescriptorToViewMap.putView(...).
Previously, ViewManager.registerView() called
MeasurementDescriptorToViewMap.getView(...) followed by
MeasurementDescriptorToViewMap.putView(...) to insert a view only if it was not
present.  There was no locking, so the view could be created multiple times.
This commit refactors putView so that it covers the full transaction on the
underlying multimap.  It creates an empty MutableView if the view is not present
and inserts it into the multimap.
