+++ core_impl/src/main/java/com/google/instrumentation/stats/MeasurementDescriptorToViewMap.java
@@ -17,11 +17,16 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import com.google.instrumentation.common.Clock;
+import com.google.instrumentation.common.Function;
+import com.google.instrumentation.stats.MutableView.MutableDistributionView;
+import com.google.instrumentation.stats.MutableView.MutableIntervalView;
+import com.google.instrumentation.stats.ViewDescriptor.DistributionViewDescriptor;
+import com.google.instrumentation.stats.ViewDescriptor.IntervalViewDescriptor;
 import java.util.Collection;
 
 /**
- * A class that stores a singleton map from {@link MeasurementDescriptor.Name}s to
- * {@link MutableView}s.
+ * A class that stores a singleton map from {@link MeasurementDescriptor.Name}s to {@link
+ * MutableView}s.
  */
 final class MeasurementDescriptorToViewMap {
 
@@ -33,12 +38,10 @@ final class MeasurementDescriptorToViewMap {
       Multimaps.synchronizedMultimap(
           HashMultimap.<MeasurementDescriptor.Name, MutableView>create());
 
-  /**
-   * Returns a {@link View} corresponding to the given {@link ViewDescriptor}.
-   */
+  /** Returns a {@link View} corresponding to the given {@link ViewDescriptor}. */
   View getView(ViewDescriptor viewDescriptor, Clock clock) {
-    Collection<MutableView> views = mutableMap.get(
-        viewDescriptor.getMeasurementDescriptor().getMeasurementDescriptorName());
+    Collection<MutableView> views =
+        mutableMap.get(viewDescriptor.getMeasurementDescriptor().getMeasurementDescriptorName());
     synchronized (mutableMap) {
       for (MutableView view : views) {
         if (view.getViewDescriptor().equals(viewDescriptor)) {
@@ -50,8 +53,8 @@ final class MeasurementDescriptorToViewMap {
   }
 
   private MutableView getMutableView(ViewDescriptor viewDescriptor) {
-    Collection<MutableView> views = mutableMap.get(
-        viewDescriptor.getMeasurementDescriptor().getMeasurementDescriptorName());
+    Collection<MutableView> views =
+        mutableMap.get(viewDescriptor.getMeasurementDescriptor().getMeasurementDescriptorName());
     synchronized (mutableMap) {
       for (MutableView view : views) {
         if (view.getViewDescriptor().equals(viewDescriptor)) {
@@ -62,11 +65,20 @@ final class MeasurementDescriptorToViewMap {
     return null;
   }
 
-  /**
-   * Map a new {@link View} to a {@link MeasurementDescriptor.Name}.
-   */
-  void putView(MeasurementDescriptor.Name name, MutableView view) {
-    mutableMap.put(name, view);
+  /** Enable stats collection for the given {@link ViewDescriptor}. */
+  void registerView(ViewDescriptor viewDescriptor, Clock clock) {
+    synchronized (mutableMap) {
+      if (getView(viewDescriptor, clock) != null) {
+        // Ignore views that are already registered.
+        return;
+      }
+      MutableView mutableView =
+          viewDescriptor.match(
+              new CreateMutableDistributionViewFunction(clock),
+              new CreateMutableIntervalViewFunction());
+      mutableMap.put(
+          viewDescriptor.getMeasurementDescriptor().getMeasurementDescriptorName(), mutableView);
+    }
   }
 
   // Records stats with a set of tags.
@@ -87,6 +99,29 @@ final class MeasurementDescriptorToViewMap {
     }
     synchronized (mutableMap) {
       view.record(tags, value);
+    }
+  }
+
+  private static final class CreateMutableDistributionViewFunction
+      implements Function<DistributionViewDescriptor, MutableView> {
+    private final Clock clock;
+
+    CreateMutableDistributionViewFunction(Clock clock) {
+      this.clock = clock;
+    }
+
+    @Override
+    public MutableView apply(DistributionViewDescriptor viewDescriptor) {
+      return MutableDistributionView.create(viewDescriptor, clock.now());
+    }
+  }
+
+  private static final class CreateMutableIntervalViewFunction
+      implements Function<IntervalViewDescriptor, MutableView> {
+    @Override
+    public MutableView apply(IntervalViewDescriptor viewDescriptor) {
+      // TODO(songya): Create Interval Aggregations from internal Distributions.
+      return MutableIntervalView.create(viewDescriptor);
     }
   }
 }
